重构

+++ RunCat365/LaunchAtStartupManager.cs
@@ -1,4 +1,4 @@
-﻿// Copyright 2025 Takuto Nakamura
+// Copyright 2025 Takuto Nakamura
 //
 //    Licensed under the Apache License, Version 2.0 (the "License");
 //    you may not use this file except in compliance with the License.
@@ -9,12 +9,10 @@
 //    Unless required by applicable law or agreed to in writing, software
 //    distributed under the License is distributed on an "AS IS" BASIS,
 //    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-    //    See the License for the specific language governing permissions and
-    //    limitations under the License.
+//    See the License for the specific language governing permissions and
+//    limitations under the License.
 
-    using System.IO;
-    using System.Reflection;
-
+using System.IO;
 
 namespace RunCat365
 {
@@ -36,11 +34,18 @@ namespace RunCat365
             {
                 if (enabled)
                 {
-                    return CreateShortcut();
+                    if (!File.Exists(_shortcutPath))
+                    {
+                        return CreateShortcut();
+                    }
+                    return true;
                 }
                 else
                 {
-                    File.Delete(_shortcutPath);
+                    if (File.Exists(_shortcutPath))
+                    {
+                        File.Delete(_shortcutPath);
+                    }
                     return true;
                 }
             }
